requirements.txt
requests==2.24.0
pandas==1.3.2
pika==1.2.0
python-dotenv